Add more tests for user registration
